refactor: consolidate ovs package into ovs package
